split helpers into: helpers.py, data_preparation.py, metrics.py and baseline_model.py
